react: Add components to manage forms
