+++ app/controllers/application_controller.rb
@@ -1,5 +1,9 @@
 class ApplicationController < ActionController::Base
-  protect_from_forgery with: :exception
+  # protect_from_forgery with: :exception
+  skip_before_action :verify_authenticity_token
+
+  helper_method :login, :logged_in?, :current_user
+
   def login(user)
     session[:id] = user.id
   end
