+++ app/controllers/users/sessions_controller.rb
@@ -1,11 +1,13 @@
 class Users::SessionsController < ApplicationController
-  skip_before_action :verify_authenticity_token, only: %i[create destroy]
+  # skip_before_action :verify_authenticity_token, only: %i[create destroy]
+  # before_action :logged_in?, except: [:destroy]
 
   def create
     @user = User.find_by(email: params[:user][:email])
     if @user && @user.authenticate(params[:user][:password])
       login @user
       render json: {
+        logged_in?: true,
         code: 200,
         user: {
           name: @user.name
@@ -14,7 +16,8 @@ class Users::SessionsController < ApplicationController
     else
       render json: {
         code: 400,
-        errors: { Error: ['Email and/or Password does not match !'] }
+        errors: { Error: ['no such user',
+                          'verify credentials and try again or signup'] }
       }
     end
   end
