+++ app/javascript/components/home/Head.jsx
