+++ app/javascript/containers/AddReading.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import Head from '../components/home/Head';
+import Head from '../components/addReading/Head';
+import ReadingForm from '../components/addReading/ReadingForm';
 import Navbar from '../components/Navbar';
-
-/* eslint-disable react/display-name */
 
 export default () => (
   <div>
     <Head />
+    <ReadingForm />
     <Navbar />
   </div>
 );
