+++ app/javascript/routes/index.jsx
@@ -6,6 +6,8 @@ import persistState from 'redux-localstorage';
 import reducer from '../reducers/sessions';
 import Readings from '../containers/Readings';
 import Home from '../containers/Home';
+import PrivateRoute from './privateRoute';
+import AddReading from '../containers/AddReading';
 
 export const store = createStore(
   reducer, compose(persistState()),
@@ -16,7 +18,8 @@ export default (
     <Router>
       <Switch>
         <Route path="/" exact component={Home} />
-        <Route path="/readings" exact component={Readings} />
+        <Route path="/addreading" component={AddReading} />
+        <PrivateRoute path="/readings" exact component={Readings} />
       </Switch>
     </Router>
   </Provider>
