+++ package.json
@@ -24,6 +24,7 @@
     "react-router-dom": "^5.2.0",
     "redux": "^4.0.5",
     "redux-localstorage": "^0.4.1",
+    "sweetalert2-react": "^0.8.3",
     "turbolinks": "^5.2.0",
     "webpack": "^4.44.1"
   },
